services/data/requirements.txt
Django==5.1.4
djangorestframework==3.15.2
django-allauth[socialaccount]==65.3.0
django-cors-headers==4.6.0
psycopg[binary]==3.2.3
gunicorn==23.0.0
whitenoise==6.8.2
